refactor(entities): move entities into their own package
update tests and dependent classes
