+++ src/main/java/brick/breaker/ICollision.java
@@ -1,5 +1,7 @@
 package brick.breaker;
 
+import brick.breaker.entities.Shape;
+
 public interface ICollision {
 
   public <T extends Shape<T>> boolean isColliding(T object);
