+++ src/main/java/brick/breaker/entities/Ball.java
@@ -1,5 +1,6 @@
-package brick.breaker;
+package brick.breaker.entities;
 
+import brick.breaker.ICollision;
 import processing.core.PApplet;
 
 public class Ball extends Shape<Ball> implements ICollision {
