+++ src/main/java/brick/breaker/entities/Brick.java
@@ -1,5 +1,6 @@
-package brick.breaker;
+package brick.breaker.entities;
 
+import brick.breaker.IHealth;
 import processing.core.PApplet;
 
 public class Brick extends Shape<Brick> implements IHealth {
