+++ src/main/java/brick/breaker/entities/Paddle.java
@@ -1,5 +1,6 @@
-package brick.breaker;
+package brick.breaker.entities;
 
+import brick.breaker.IMovement;
 import processing.core.PApplet;
 import processing.core.PVector;
